+++ components/common/CharmEditor/components/@bangle.dev/core/plugin-loader.ts
@@ -1,7 +1,7 @@
 // References: https://github.com/bangle-io/bangle-editor/blob/13127cf2e4187ebaa6d5e01d80f4e9018fae02a5/lib/core/src/plugin-loader.ts
 
 import { log } from '@charmverse/core/log';
-import * as history from '@packages/charmeditor/plugins/history';
+import * as history from '@packages/charmeditor/extensions/history';
 import { baseKeymap as pmBaseKeymap } from 'prosemirror-commands';
 import { gapCursor as pmGapCursor } from 'prosemirror-gapcursor';
 import { InputRule, inputRules as pmInputRules, undoInputRule as pmUndoInputRule } from 'prosemirror-inputrules';
