+++ packages/charmeditor/src/extensions/listItem/consolidateListNodes.ts
