+++ packages/charmeditor/src/extensions/listItem/listItemNodeView.ts
