+++ packages/charmeditor/src/extensions/listItem/mergeListItems.ts
