+++ packages/charmeditor/src/schema.ts
@@ -1,6 +1,6 @@
 import { buildSchema } from './buildSchema';
-import { bulletList } from './specs/bulletList';
-import { listItem } from './specs/listItem';
-import { tabIndentSpec } from './specs/tabIndent';
+import { bulletList } from './extensions/bulletList';
+import { listItem } from './extensions/listItem/listItemSpec';
+import { spec as tabIndentSpec } from './extensions/tabIndent';
 
 export const schema = buildSchema([bulletList, listItem, tabIndentSpec]);
